// src/common/constants/token-types.enum.ts
export enum TOKEN_TYPE {
  LOGIN = 'LOGIN_TOKEN',
  ADMIN_LOGIN = 'ADMIN_LOGIN_TOKEN',
  CUSTOMER_LOGIN = 'CUSTOMER_LOGIN_TOKEN',
  TRUCK_OWNER_LOGIN = 'TRUCK_OWNER_LOGIN_TOKEN',
  DRIVER_LOGIN = 'DRIVER_LOGIN_TOKEN',
  REGISTER = 'REGISTER_TOKEN',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD_TOKEN',
  CHANGE_EMAIL = 'CHANGE_EMAIL_TOKEN',
  RESET_PASSWORD = 'RESET_PASSWORD',
  SET_PASSWORD = 'SET_PASSWORD',

  VERIFY = 'VERIFY',
  PUBLIC_TOKEN = 'PUBLIC_TOKEN',
}