// src/entities/order/enums/order-status.enum.ts
export enum ORDER_STATUS {
  CREATED = 'CREATED',
  VERIFIED = 'VERIFIED',
  ASSIGNING = 'ASSIGNING',
  ASSIGNED = 'ASSIGNED',
  DISPATCHED = 'DISPATCHED',
  PICKING = 'PICKING',
  PICK_ARRIVED = 'PICK_ARRIVED',
  PICKUPCODE_INPUTED = 'PICKUPCODE_INPUTED',
  DELIVERING = 'DELIVERING',
  DELIVERYCODE_INPUTED = 'DELIVERYCODE_INPUTED',
  DELIVERED = 'DELIVERED',
  CANCELING = 'CANCELING',
  CANCELED = 'CANCELED',
  CUSTCANCEL = 'CUSTCANCEL',
  DRIVERCANCEL = 'DRIVERCANCEL',
}